# Fix `pattern AndBool` and friends that did not match the `_andBool_` symbol