--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -13,7 +13,7 @@ concurrency:
 
 env:
   ghc_version: '9.2.8'
-  hlint_version: '3.4.1'
+  hlint_version: '3.5'
   fourmolu_version: '0.12.0.0'
   hpack_version: '0.34.2'