--- a/test/predicates-integration/Main.hs
+++ b/test/predicates-integration/Main.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE PatternSynonyms #-}
-
 {- |
 Copyright   : (c) Runtime Verification, 2023
 License     : BSD-3-Clause