--- a/test/predicates-integration/definition/predicates.k
+++ b/test/predicates-integration/definition/predicates.k
@@ -1,0 +1,6 @@
+module PREDICATES
+  imports INT
+  imports BYTES
+  imports K-EQUAL
+
+endmodule