… package.yaml
… test/predicates-integration/README.md