Summary of changes:
* correct the pattern synonyms in `Booster.Pattern.Bool` by including the `hook` attribute to the hooked symbols
  - the absence of the attribute in i.e. `AndBool` pattern synonym caused it to *not match* the `_andBool_` symbols from rules's requires clauses;
  - that caused the conjunction splitting functions `splitAndBool` and `splitBoolPredicates` to have no effect, potentially blocking rule applications.
* add unit-tests for these pattern synonyms in `Test.Booster.Pattern.Bool`
  - the test internalize the symbols from surface KORE using the quasi-quoter and test that the internalized symbols are indeed matched by the pattern synonyms.
* change the hook name to be represented as `ByteString` (not `Text`)
  - having both BS and Text causes issues when parsing symbols with both hook and smt-hook attributes using the quasi-quoter
  - that triggered the need to regenerate the responses in `test-foundry-bug-report`

TODO:

- [x] run KEVM performance tests and post results
- [x] run Kontrol performance tests and post results