@geo2a 2024-01-25
**KEVM performance**: no miracle...
| Test                                    | georgy-better-unclear-condition time | master-a956edb time | (georgy-better-unclear-condition/master-a956edb) time
|-----------------------------------------|--------------------------------------|---------------------|-------------------------------------------------------
| mcd/flipper-addu48u48-fail-rough-spec.k | 98.09                                | 111.61              | 0.8788639010841323
| erc20/hkg/approve-spec.k                | 110.39                               | 123.05              | 0.8971149939049168
| erc20/hkg/transferFrom-failure-2-spec.k | 241.72                               | 263.33              | 0.9179356700717731
| mcd/flopper-cage-pass-spec.k            | 178.48                               | 193.87              | 0.9206169082374787
| mcd/vat-mului-pass-spec.k               | 272.78                               | 288.6               | 0.9451836451836451
| mcd/vat-subui-pass-rough-spec.k         | 292.52                               | 307.96              | 0.9498636186517729
| mcd/vat-subui-fail-rough-spec.k         | 148.88                               | 156.27              | 0.9527100531132014
| mcd/cat-file-addr-pass-rough-spec.k     | 190.1                                | 198.48              | 0.9577791213220476
| erc20/hkg/transfer-failure-2-spec.k     | 133.85                               | 139.61              | 0.9587422104433778
| mcd/vat-muluu-pass-spec.k               | 255.08                               | 241.48              | 1.0563193639224782
| benchmarks/storagevar00-spec.k          | 86.91                                | 82.22               | 1.0570420822184383
| mcd/vat-subui-pass-spec.k               | 311.04                               | 290.93              | 1.069123156773107
| mcd/end-subuu-pass-spec.k               | 247.93                               | 230.56              | 1.0753383067314366
| erc20/hkg/balanceOf-spec.k              | 111.47                               | 100.53              | 1.1088232368447228
| TOTAL                                   | 2679.24                              | 2728.5              | 0.9819461242440901