@geo2a 2024-01-31
# Kontrol performance

| Test                                                                   | georgy-better-unclear-condition time | master-0ef6ecd time | (georgy-better-unclear-condition/master-0ef6ecd) time
|------------------------------------------------------------------------|--------------------------------------|---------------------|-------------------------------------------------------
| AssumeTest.testFail_assume_true(uint256,uint256)                       | 68.82                                | 102.85              | 0.6691298006806028
| AssumeTest.test_assume_false(uint256,uint256)                          | 75.75                                | 108.51              | 0.6980923417196572
| src/tests/integration/test_foundry_prove.py::test_foundry_resume_proof | 76.34                                | 108.96              | 0.7006240822320118
| ExpectRevertTest.testFail_expectRevert_multipleReverts()               | 171.95                               | 194.52              | 0.8839707999177462
| EmitContractTest.testExpectEmit()                                      | 131.91                               | 140.62              | 0.9380600199118191
| BMCLoopsTest.test_countdown_concrete()                                 | 24.55                                | 26.16               | 0.9384556574923547
| BytesTypeTest.test_bytes32(bytes32)                                    | 25.22                                | 26.33               | 0.9578427649069503
| src/tests/integration/test_foundry_prove.py::test_foundry_remove_node  | 44.76                                | 43.15               | 1.0373117033603707
| AssertTest.testFail_assert_true()                                      | 39.66                                | 37.87               | 1.0472669659360971
| CounterTest.testSetNumber(uint256)                                     | 247.49                               | 235.03              | 1.0530145087861125
| TOTAL                                                                  | 906.4499999999999                    | 1024.0              | 0.8852050781249999