@geo2a 2024-02-01
I've re-run KEVM performance again since both `main` and this branch had changes: 
```
| Test                                     | georgy-better-unclear-condition time | master-0ef6ecd time | (georgy-better-unclear-condition/master-0ef6ecd) time
|------------------------------------------|--------------------------------------|---------------------|-------------------------------------------------------
| erc20/hkg/transferFrom-success-2-spec.k  | 31.67                                | 40.91               | 0.7741383524810561
| benchmarks/storagevar01-spec.k           | 27.12                                | 28.35               | 0.9566137566137566
| erc20/hkg/transfer-failure-2-spec.k      | 27.33                                | 28.4                | 0.9623239436619718
| mcd/dsvalue-read-pass-spec.k             | 72.96                                | 75.78               | 0.962787015043547
| erc20/ds/transferFrom-failure-2-c-spec.k | 28.27                                | 29.34               | 0.9635310156782549
| erc20/hkg/transfer-success-1-spec.k      | 27.38                                | 28.39               | 0.9644240929904896
| erc20/hkg/transfer-failure-1-spec.k      | 26.9                                 | 27.89               | 0.9645034062387952
| benchmarks/requires01-a0gt0-spec.k       | 27.41                                | 28.41               | 0.9648011263639563
| mcd/vat-addui-fail-rough-spec.k          | 103.96                               | 107.69              | 0.9653635435045036
| benchmarks/storagevar03-spec.k           | 27.35                                | 28.32               | 0.9657485875706215
| erc20/hkg/transferFrom-failure-1-spec.k  | 33.22                                | 30.51               | 1.0888233366109472
| bihu/forwardToHotWallet-failure-1-spec.k | 119.25                               | 107.98              | 1.10437117984812
| TOTAL                                    | 552.82                               | 561.97              | 0.9837179920636333
```